\name{iecs-package}
\alias{iecs-package}
\docType{package}
\title{
Iterative Exclusion of Compatible Samples Analysis of the Data
}
\description{
pkg{iecs} is a package to perform the Iterative Exclusion of Compatible Samples(iecs) Analysis and Necessary Conditons Analysis of data.
}
\author{

\bold{Author}:\cr
Wei Xu\cr

\bold{Maintainer}:\cr
Wei Xu <eric_wxu@126.com>\cr
}
\references{
Charles C. Ragin.The Comparative Method. Moving Beyond Qualitative and Quantitative Strategies, University of California Press, Berkeley and Los Angeles (1987)
}

